vectorstore-in-memory/pyproject.toml
[tool.poetry]
name = "vectorstore-in-memory"
version = "0.1.0"
description = ""
authors = ["Erick Cardenas <erick.mendez@toptal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.6"
python-dotenv = "^1.0.1"
langchain-openai = "^0.0.5"
pypdf = "^4.0.1"
faiss-cpu = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
